move get_ecr_login to ecr module and patch in hokusai config fixture
