+++ test/fixtures/config.yml
@@ -1,0 +1,4 @@
+---
+aws-account-id: '12345'
+aws-ecr-region: us-east-1
+project-name: foo
